functions/generate-embeddings/requirements.txt
functions-framework==3.4.0
google-cloud-bigquery==3.4.0
google-cloud-aiplatform==1.38.0
google-auth==2.23.0
requests==2.31.0
beautifulsoup4==4.12.0
html5lib==1.1
lxml==4.9.3
pandas==1.5.3
numpy==1.24.3
